# edge/medbook-bff/src/main/resources/openapi/bff-api_v1.0.yaml
openapi: 3.0.3
...
info:
  title: MedBook - BFF API
  description: |
    ## MedBook BFF
    Unico punto di ingresso del frontend verso il sistema.
    Orchestra le chiamate ai DMN via Feign client e restituisce risposte aggregate.
    Non ha database proprio. E' un orchestratore stateless.

    ### Ruoli supportati
    - ROLE_PATIENT: paziente registrato
    - ROLE_DOCTOR: medico
    - ROLE_RECEPTIONIST: addetto alla reception
    - ROLE_ADMIN: amministratore

  version: 1.0.0
  contact:
    name: MedBook Team
    email: djizalfra@gmail.com
servers:
  - url: http://localhost:8090
    description: BFF locale - sviluppo
...
paths:
...
  /bff/v1/patients:

    post:
      tags:
        - Patients
      operationId: postCreatePatient
      summary: Registra un nuovo paziente
      parameters:
        - $ref: '#/components/parameters/MedBookContextParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePatientBffRequest'
      responses:
        '201':
          description: Paziente registrato con successo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MedBookApiResponse'
        '422':
          description: Dati non validi
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MedBookApiErrorResponse'

    get:
      tags:
        - Patients
      operationId: getAllPatients
      summary: Lista pazienti con filtri
      parameters:
        - $ref: '#/components/parameters/MedBookContextParam'
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - $ref: '#/components/parameters/SortParam'
        - name: status
          in: query
          required: false
          schema:
            type: string
        - name: lastName
          in: query
          required: false
          schema:
            type: string
        - name: city
          in: query
          required: false
          schema:
            type: string
        - name: email
          in: query
          required: false
          schema:
            type: string
        - name: firstName
          in: query
          required: false
          schema:
            type: string
        - name: fiscalCode
          in: query
          required: false
          schema:
            type: string
        - name: phone
          in: query
          required: false
          schema:
            type: string
        - name: gender
          in: query
          required: false
          schema:
            type: string
        - name: province
          in: query
          required: false
          schema:
            type: string
        - name: createdFrom
          in: query
          required: false
          schema:
            type: string
            format: date
        - name: createdTo
          in: query
          required: false
          schema:
            type: string
            format: date
        - name: updatedFrom
          in: query
          required: false
          schema:
            type: string
            format: date
        - name: updatedTo
          in: query
          required: false
          schema:
            type: string
            format: date
      responses:
        '200':
          description: Lista pazienti recuperata con successo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MedBookApiResponse'
  /bff/v1/patients/me:

    get:
      tags:
        - Patients
      operationId: getMyPatient
      summary: Profilo paziente dell'utente autenticato
      description: |
        Restituisce il profilo del paziente autenticato.
        Il patientId viene risolto internamente cercando per lo username (CF per pazienti, email per medici) del claim
        'preferred_username' nel token JWT - non e' necessario passarlo come parametro.
      parameters:
        - $ref: '#/components/parameters/MedBookContextParam'
      responses:
        '200':
          description: Profilo paziente trovato
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MedBookApiResponse'
        '404':
          description: Profilo paziente non trovato per l'utente autenticato
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MedBookApiErrorResponse'

    patch:
      tags:
        - Patients
      operationId: patchMyPatient
      summary: Aggiorna il profilo paziente dell'utente autenticato
      description: |
        Aggiorna i dati modificabili del profilo paziente autenticato.
        Il patientId viene risolto internamente - non e' necessario passarlo come parametro.
      parameters:
        - $ref: '#/components/parameters/MedBookContextParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePatientBffRequest'
      responses:
        '200':
          description: Profilo aggiornato con successo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MedBookApiVoidResponse'
        '404':
          description: Profilo paziente non trovato per l'utente autenticato
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MedBookApiErrorResponse'
  /bff/v1/patients/{patientId}:

    get:
      tags:
        - Patients
      operationId: getPatientById
      summary: Dettaglio paziente
      parameters:
        - $ref: '#/components/parameters/MedBookContextParam'
        - $ref: '#/components/parameters/PatientIdPathParam'
      responses:
        '200':
          description: Paziente trovato
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MedBookApiResponse'
        '404':
          description: Paziente non trovato
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MedBookApiErrorResponse'

    patch:
      tags:
        - Patients
      operationId: patchUpdatePatient
      summary: Aggiorna dati paziente
      parameters:
        - $ref: '#/components/parameters/MedBookContextParam'
        - $ref: '#/components/parameters/PatientIdPathParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdatePatientBffRequest'
      responses:
        '200':
          description: Paziente aggiornato con successo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MedBookApiVoidResponse'
        '404':
          description: Paziente non trovato
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MedBookApiErrorResponse'

    delete:
      tags:
        - Patients
      operationId: deletePatient
      summary: Elimina logicamente un paziente
      parameters:
        - $ref: '#/components/parameters/MedBookContextParam'
        - $ref: '#/components/parameters/PatientIdPathParam'
      responses:
        '200':
          description: Paziente eliminato con successo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MedBookApiVoidResponse'
        '404':
          description: Paziente non trovato
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MedBookApiErrorResponse'
  /bff/v1/patients/{patientId}/restore:

    patch:
      tags:
        - Patients
      operationId: patchRestorePatient
      summary: Ripristina un paziente eliminato logicamente
      parameters:
        - $ref: '#/components/parameters/MedBookContextParam'
        - $ref: '#/components/parameters/PatientIdPathParam'
      responses:
        '200':
          description: Paziente ripristinato con successo
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MedBookApiVoidResponse'
        '404':
          description: Paziente non trovato
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MedBookApiErrorResponse'
        '409':
          description: Il paziente non e cancellato - impossibile ripristinare
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MedBookApiErrorResponse'

  # ===========================================================================
  # DOCTORS
  # ===========================================================================
...
components:
...
  parameters:
    MedBookContextParam:
      name: X-MedBook-Context
      in: header
      required: false
      description: |
        Contesto applicativo della richiesta - iniettato da MedBookContextInterceptor.
        Non viene letto dall'header HTTP: il resolver preleva MedBookContextHolder.get().
      schema:
        $ref: '#/components/schemas/MedBookContext'
    PatientIdPathParam:
      name: patientId
      in: path
      required: true
      schema:
        type: string
        example: PAT-1
...
    PageParam:
      name: page
      in: query
      required: false
      schema:
        type: integer
        default: 0
        minimum: 0
    SizeParam:
      name: size
      in: query
      required: false
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
    SortParam:
      name: sort
      in: query
      required: false
      schema:
        type: string
        example: createdAt,desc

  # ---------------------------------------------------------------------------
  # SCHEMAS
  # ---------------------------------------------------------------------------
  schemas:
...
    MedBookApiResponse:
      type: object
      description: Risposta di successo con payload.
      properties:
        httpStatus:
          type: integer
        success:
          type: boolean
        timestamp:
          type: string
          format: date-time
        traceId:
          type: string
        data:
          description: Payload della risposta
        page:
          $ref: '#/components/schemas/MedBookPageResponse'
    MedBookApiVoidResponse:
      type: object
      description: Risposta base per endpoint senza payload.
      properties:
        httpStatus:
          type: integer
        success:
          type: boolean
        timestamp:
          type: string
          format: date-time
        traceId:
          type: string
    MedBookApiErrorResponse:
      type: object
      description: Formato standard delle risposte di errore.
      properties:
        httpStatus:
          type: integer
        success:
          type: boolean
        timestamp:
          type: string
          format: date-time
        traceId:
          type: string
        errorCode:
          type: string
        message:
          description: Messaggio di errore
    MedBookPageResponse:
      type: object
      description: Metadati di paginazione.
      properties:
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
        size:
          type: integer
        number:
          type: integer
        first:
          type: boolean
        last:
          type: boolean
        empty:
          type: boolean
    MedBookContext:
      type: object
      description: Contesto della richiesta HTTP.
      properties:
        traceId:
          type: string
        username:
          type: string
        roles:
          type: array
          items:
            type: string
        requestTimestamp:
          type: string
          format: date-time

    # =========================================================================
    # PATIENT REQUEST
    # =========================================================================
    CreatePatientBffRequest:
      type: object
      description: Dati per la registrazione di un nuovo paziente
      required:
        - firstName
        - lastName
        - dateOfBirth
        - gender
        - email
        - fiscalCode
        - consensoPrivacy
      properties:
        firstName:
          type: string
          example: Mario
        lastName:
          type: string
          example: Rossi
        dateOfBirth:
          type: string
          format: date
          example: "1990-01-15"
        fiscalCode:
          type: string
          example: RSSMRA90A15H501Z
        gender:
          type: string
          example: MASCHILE
        email:
          type: string
          format: email
          example: mario.rossi@email.it
        phone:
          type: string
          example: "+393331234567"
        address:
          type: string
          example: Via Roma 12
        city:
          type: string
          example: Napoli
        postalCode:
          type: string
          example: "80100"
        province:
          type: string
          maxLength: 100
          description: Sigla provincia (2 caratteri, es. TO, MI, NA)
          example: TO
        comuneNascita:
          type: string
          maxLength: 100
          description: Comune di nascita (o citta estera per nati fuori Italia)
        provinciaNascita:
          type: string
          maxLength: 100
          description: Sigla provincia di nascita (o 'EE' per nati all'estero)
          example: TO
        regioneNascita:
          type: string
          maxLength: 100
          description: Regione di nascita (o stato estero)
        consensoPrivacy:
          type: boolean
          description: Consenso obbligatorio alla privacy policy e termini di utilizzo
        consensoCommerciale:
          type: boolean
          description: Consenso facoltativo a comunicazioni commerciali
          default: false
        consensoProfilazione:
          type: boolean
          description: Consenso facoltativo alla profilazione
          default: false
        emailEnabled:
          type: boolean
          description: true se il canale EMAIL è abilitato
          example: true
        smsEnabled:
          type: boolean
          description: true se il canale SMS è abilitato
          example: false
        password:
          type: string
          format: password
          description: |
            Password dell'account Keycloak. Presente solo durante la registrazione
            pubblica del paziente. Non viene mai loggata ne' salvata nel DB MedBook.
          minLength: 8
          example: Password1!
    UpdatePatientBffRequest:
      type: object
      description: Dati per l'aggiornamento parziale di un paziente
      properties:
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
          format: email
        phone:
          type: string
        address:
          type: string
        city:
          type: string
        postalCode:
          type: string
        province:
          type: string
          maxLength: 100
          description: Sigla provincia (2 caratteri, es. TO, MI, NA)
          example: TO
        comuneNascita:
          type: string
          description: Comune di nascita (o citta estera per nati fuori Italia)
        provinciaNascita:
          type: string
          maxLength: 100
          description: Sigla provincia di nascita (o 'EE' per nati all'estero)
          example: TO
        regioneNascita:
          type: string
          description: Regione di nascita (o stato estero)
        consensoCommerciale:
          type: boolean
          description: Consenso facoltativo a comunicazioni commerciali
        consensoProfilazione:
          type: boolean
          description: Consenso facoltativo alla profilazione
        status:
          type: string

    # =========================================================================
    # DOCTOR REQUEST
    # =========================================================================
...